+++ path-tracker-pipelines/src/main/java/org/example/Main.java
@@ -18,21 +18,19 @@
 
 package org.example;
 
-import org.apache.flink.streaming.api.environment.PathAnalyzer;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.kafka.clients.admin.AdminClient;
-import org.apache.kafka.clients.admin.AdminClientConfig;
-import org.apache.kafka.clients.admin.NewTopic;
+
+import org.example.merger.KafkaConsumerThread;
+import org.example.merger.KafkaMergeThread;
+import org.example.merger.kafkaMessage;
 import org.example.pipelines.ConfluxPipeline;
 import org.example.pipelines.GlobalSortPipeline;
 import org.example.utils.KafkaAdminUtils;
 import org.testcontainers.containers.KafkaContainer;
 import org.testcontainers.utility.DockerImageName;
-import java.util.Collections;
-import java.util.Properties;
+
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.atomic.AtomicLong;
 
 public class Main {
